+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2019-02-12
+
+#### Added
+* support for [Filter Bundle](https://github.com/heimrichhannot/contao-filter-bundle) AdjustFilterOptions event
+
+#### Changed
+* updated js file name to new internal conventions
+* updated refactored FlatpickrBundle object to class
+
 ## [0.1.3] - 2019-01-30
 
 #### Fixed